tabled/ Move ByColumnName to its own file
